statsv: Consolidate verbose/dry-run and improve arg docs
There was only 1 message gated by --verbose, which is the statsd
output, intended for use with --dry-run.

Instead of having --dry-run confusingly be silent by default,
have --dry-run log the statsd message always. After this, there is
no use of --verbose left, so it can be removed.

This is a no-op for WMF production, as neither parameters are used
in Puppet.

While at it, re-order the argument definitions such that the Kafka
configuration is all grouped together, with the descriptions clearly
mentioning Kafka at the start.

Change-Id: I3fa095b06c2d058001136d0c15baa7a09531dda8