Implement `--kafka-fixture` and basic integration test
== Alternatives considered ==

In previous commits, I factored out statsv_main for potential use in
a Python unittest file, but went with testing from Bash for now.

Using Python's built-in unittest, I set up a DGRAM socket (acting as
statsd receiver) on an available port (e.g. sock.bind 0, and
getsockname[1]), then call statsv_main with `--kafka-fixture x.txt`
`--statsd localhost:$PORT` to fully test it end-to-end (besides Kafka).
I got stock finding a way to make this work without dependencies and
in a single process, given that most socket interaction in Python
is blocking, so there isn't an obvious way for socket.recv (in test)
and socket.sendto (in source) to take turns, without a lot of boiler
plate. In addition, `import statsv` to reference a relative file is
deprecated in Python 3.11:
```
ImportWarning: _SixMetaPathImporter.find_spec() not found; falling back to find_module()
```

Using `from .statsv` is unsupported outside a package, so we'd have to
first move code to `statsv/__init__.py` and possibly coordinate this
with Puppet. Easy to do, but given the test didn't work out anyway,
left as-is for now.

Change-Id: Ie4d9ecdd51d75b7cd45fe1a9a2dd4b6240c663d2